requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
Flask-CORS==5.0.0
psycopg2-binary==2.9.10
PyJWT==2.10.1
bcrypt==4.2.1
python-dotenv==1.0.1
